+++ components/commercetools-ui/atoms/overlay/index.tsx
@@ -7,8 +7,7 @@ interface Props {
 }
 
 const Overlay: React.FC<Props> = ({ zIndex, onClick }) => {
-  //eslint-disable-next-line tailwindcss/no-custom-classname
-  const z = `z-[${zIndex ?? 600}]`;
+  const z = `z-[${zIndex ?? 300}]`;
   return (
     <div className={useClassNames(['fixed left-0 top-0 h-screen w-screen bg-[#7f7f7f]/30', z])} onClick={onClick} />
   );