+++ providers/index.tsx
@@ -8,24 +8,33 @@ import { sdk } from 'sdk';
 import { Translations } from 'types/i18n';
 import I18nProvider from './i18n';
 import { SWRProvider } from './swr';
+import TracingProvider from './tracing';
 import { GetAccountActionReturn } from '../sdk/composable-commerce/types/actions/AccountActions';
 
 interface ProvidersProps {
   translations: Translations;
   accountResult?: SDKResponse<GetAccountActionReturn>;
+  tracing: SDKResponse<unknown>['tracing'];
 }
 
-export const Providers = ({ translations, accountResult, children }: React.PropsWithChildren<ProvidersProps>) => {
+export const Providers = ({
+  translations,
+  accountResult,
+  tracing,
+  children,
+}: React.PropsWithChildren<ProvidersProps>) => {
   const { locale } = useParams();
 
   sdk.defaultConfigure(locale);
 
   return (
-    <I18nProvider translations={translations}>
-      <SWRProvider value={{ fallback: { '/action/account/getAccount': accountResult } }}>
-        <AddToCartOverlayProvider>{children}</AddToCartOverlayProvider>
-        <Toaster />
-      </SWRProvider>
-    </I18nProvider>
+    <TracingProvider tracing={tracing}>
+      <I18nProvider translations={translations}>
+        <SWRProvider value={{ fallback: { '/action/account/getAccount': accountResult } }}>
+          <AddToCartOverlayProvider>{children}</AddToCartOverlayProvider>
+          <Toaster />
+        </SWRProvider>
+      </I18nProvider>
+    </TracingProvider>
   );
 };