+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+## Version 1.11.4 (2024-04-26)
+
+** New Feaures and Improvements **
+
+- API Request Id is logged in windows console
+- Search overlay fixed when side panel is open
+
 ## Version 1.11.3 (2024-04-19)
 
 ** New Features and Improvements **